refactor: extract configuration from server.ts and api.ts into dedicated config files
- packages/gitea-client/tokenScopes.ts: shared source of truth for
  resolveGiteaTokenScopes, eliminating duplication between server.ts and
  apps/app/giteaTokenScopes.ts
- services/api/config.ts: all env var reads, parse utilities, startup
  validation, and derived values (allowedOrigins, tokenScopes, cookie
  settings) extracted from the top ~150 lines of server.ts
- apps/app/config.ts: apiBaseUrl extracted from api.ts

server.ts now starts directly with handler utilities; api.ts has no
top-level config. All 167 tests pass.
