+++ services/api/runtime-auth.test.ts
@@ -2,10 +2,12 @@ import { describe, expect, test } from "bun:test";
 import { readFileSync } from "node:fs";
 
 const serverSource = readFileSync("services/api/server.ts", "utf8");
+const configSource = readFileSync("services/api/config.ts", "utf8");
 
 describe("API runtime Gitea auth", () => {
   test("uses the dedicated service token for production Gitea calls", () => {
-    expect(serverSource).toContain("BINDERSNAP_GITEA_SERVICE_TOKEN");
+    // BINDERSNAP_GITEA_SERVICE_TOKEN is read in config.ts (the config extraction module)
+    expect(configSource).toContain("BINDERSNAP_GITEA_SERVICE_TOKEN");
     // All privileged call sites must go through the wrapper, not directly to the
     // service-token helper, so the dev fallback is applied consistently.
     expect(serverSource).toContain("buildGiteaPrivilegedHeaders");
