+++ services/api/server.ts
@@ -1,6 +1,33 @@
 import { randomUUID } from "crypto";
 
 import { logger } from "./logger";
+import {
+  apiPort,
+  giteaUrl,
+  giteaAdminUsername,
+  giteaAdminPassword,
+  giteaServiceToken,
+  isProduction,
+  stripeSecretKey,
+  stripeWebhookSecret,
+  stripePriceId,
+  appOrigin,
+  sessionCookieName,
+  tokenScopes,
+  sessionTtlMs,
+  rememberedSessionTtlMs,
+  enforceHttps,
+  authRateLimitEnabled,
+  bypassSubscriptionForUsers,
+  authRateLimitWindowMs,
+  authRateLimitMax,
+  sessionCookieDomain,
+  sessionCookieSameSite,
+  allowedOrigins,
+  normalizeOrigin,
+  resolveCookieSameSite,
+  type SessionCookieSameSite,
+} from "./config";
 import { sessionStore, type SessionRecord } from "./sessions";
 import { subscriptionStore, hasActiveSubscription } from "./subscriptions";
 import { verifyStripeSignature } from "./stripe/webhook";
@@ -44,185 +71,10 @@ import {
   type PullRequestWithApprovalState,
 } from "../../packages/gitea-client/pullRequests";
 
-function parseBoolean(value: string | undefined, fallback: boolean): boolean {
-  if (value === undefined) {
-    return fallback;
-  }
-
-  const normalized = value.trim().toLowerCase();
-  if (["1", "true", "yes", "on"].includes(normalized)) {
-    return true;
-  }
-
-  if (["0", "false", "no", "off"].includes(normalized)) {
-    return false;
-  }
-
-  return fallback;
-}
-
-function parsePositiveInt(value: string | undefined, fallback: number): number {
-  const parsed = Number.parseInt(value ?? "", 10);
-  return Number.isFinite(parsed) && parsed > 0 ? parsed : fallback;
-}
-
-const apiPortValue = Number.parseInt(
-  process.env.API_PORT ?? process.env.PORT ?? "8787",
-  10,
-);
-const apiPort =
-  Number.isFinite(apiPortValue) && apiPortValue > 0 ? apiPortValue : 8787;
-const giteaUrl =
-  process.env.GITEA_INTERNAL_URL ??
-  process.env.BUN_PUBLIC_GITEA_URL ??
-  process.env.VITE_GITEA_URL ??
-  "http://localhost:3000";
-const giteaAdminUsername = process.env.GITEA_ADMIN_USER?.trim() ?? "";
-const giteaAdminPassword = process.env.GITEA_ADMIN_PASS?.trim() ?? "";
-const giteaServiceToken =
-  process.env.BINDERSNAP_GITEA_SERVICE_TOKEN?.trim() ?? "";
-const isProduction = process.env.NODE_ENV === "production";
-
-if (isProduction && !giteaServiceToken) {
-  logger.error(
-    "FATAL: BINDERSNAP_GITEA_SERVICE_TOKEN is not set in production",
-    { env: "production" },
-  );
-  process.exit(1);
-}
-
-const stripeSecretKey = process.env.STRIPE_SECRET_KEY?.trim() ?? "";
-const stripeWebhookSecret = process.env.STRIPE_WEBHOOK_SECRET?.trim() ?? "";
-const stripePriceId = process.env.STRIPE_PRICE_ID?.trim() ?? "";
-
-const defaultStripeAppOrigin = "http://localhost:5173";
-const appOrigin = (
-  process.env.BINDERSNAP_APP_ORIGIN ??
-  process.env.BINDERSNAP_ALLOWED_ORIGINS?.split(",")[0] ??
-  defaultStripeAppOrigin
-).trim();
-
-if (isProduction && !stripeSecretKey) {
-  logger.error("FATAL: STRIPE_SECRET_KEY is not set in production", {
-    env: "production",
-  });
-  process.exit(1);
-}
-
-if (isProduction && !stripeWebhookSecret) {
-  logger.error("FATAL: STRIPE_WEBHOOK_SECRET is not set in production", {
-    env: "production",
-  });
-  process.exit(1);
-}
-
-const emailDomain =
-  process.env.BINDERSNAP_USER_EMAIL_DOMAIN ?? "users.bindersnap.local";
-const sessionCookieName =
-  process.env.BINDERSNAP_SESSION_COOKIE_NAME ?? "bindersnap_session";
-export type SessionCookieSameSite = "Strict" | "Lax" | "None";
-const REQUIRED_GITEA_TOKEN_SCOPES = [
-  "write:user",
-  "write:repository",
-  "write:issue",
-] as const;
-
-function resolveGiteaTokenScopes(scopesRaw?: string): string[] {
-  const configuredScopes = (scopesRaw ?? "")
-    .split(",")
-    .map((scope) => scope.trim())
-    .filter((scope) => scope !== "");
-
-  return Array.from(
-    new Set<string>([...configuredScopes, ...REQUIRED_GITEA_TOKEN_SCOPES]),
-  );
-}
-
-const tokenScopes = resolveGiteaTokenScopes(
-  process.env.BINDERSNAP_GITEA_TOKEN_SCOPES,
-);
-const sessionTtlMs = parsePositiveInt(
-  process.env.BINDERSNAP_SESSION_TTL_MS,
-  7 * 24 * 60 * 60 * 1000,
-);
-const rememberedSessionTtlMs = parsePositiveInt(
-  process.env.BINDERSNAP_REMEMBER_ME_SESSION_TTL_MS,
-  30 * 24 * 60 * 60 * 1000,
-);
-const enforceHttps = parseBoolean(
-  process.env.BINDERSNAP_REQUIRE_HTTPS,
-  process.env.NODE_ENV === "production",
-);
-const authRateLimitEnabled = parseBoolean(
-  process.env.BINDERSNAP_AUTH_RATE_LIMIT_ENABLED,
-  true,
-);
-const bypassSubscriptionForUsers = new Set(
-  (process.env.BINDERSNAP_FREE_USERS ?? "")
-    .split(",")
-    .map((u) => u.trim())
-    .filter((u) => u !== ""),
-);
-const authRateLimitWindowMs = parsePositiveInt(
-  process.env.BINDERSNAP_AUTH_RATE_LIMIT_WINDOW_MS,
-  10 * 60 * 1000,
-);
-const authRateLimitMax = parsePositiveInt(
-  process.env.BINDERSNAP_AUTH_RATE_LIMIT_MAX,
-  20,
-);
-const defaultAppOrigin = `http://localhost:${process.env.APP_PORT ?? "5173"}`;
-const configuredAllowedOrigins = (
-  process.env.BINDERSNAP_ALLOWED_ORIGINS ??
-  process.env.BINDERSNAP_APP_ORIGIN ??
-  defaultAppOrigin
-)
-  .split(",")
-  .map((origin) => origin.trim())
-  .filter((origin) => origin !== "");
 
 const authAttempts = new Map<string, { count: number; resetAt: number }>();
-const sessionCookieDomain = resolveCookieDomain(
-  process.env.BINDERSNAP_SESSION_COOKIE_DOMAIN,
-);
-const sessionCookieSameSite = resolveCookieSameSite(
-  process.env.BINDERSNAP_SESSION_COOKIE_SAME_SITE,
-);
-const allowedOrigins = new Set(
-  configuredAllowedOrigins
-    .map(normalizeOrigin)
-    .filter((origin): origin is string => Boolean(origin)),
-);
 
-function resolveCookieDomain(value: string | undefined): string | null {
-  if (!value) {
-    return null;
-  }
-
-  const trimmed = value.trim();
-  if (trimmed === "") {
-    return null;
-  }
-
-  return /^\.?[a-z0-9.-]+$/i.test(trimmed) ? trimmed : null;
-}
-
-export function resolveCookieSameSite(
-  value: string | undefined,
-  fallback: SessionCookieSameSite = "Lax",
-): SessionCookieSameSite {
-  const normalized = value?.trim().toLowerCase();
-  switch (normalized) {
-    case "strict":
-      return "Strict";
-    case "none":
-      return "None";
-    case "lax":
-      return "Lax";
-    default:
-      return fallback;
-  }
-}
+export { resolveCookieSameSite, type SessionCookieSameSite };
 
 export interface SessionLifetime {
   sessionExpiresAt: number;
@@ -246,23 +98,6 @@ export function buildSessionLifetime(
   return rememberMe
     ? { sessionExpiresAt: expiresAt, cookieExpiresAt: expiresAt }
     : { sessionExpiresAt: expiresAt };
-}
-
-function normalizeOrigin(origin: string | null | undefined): string | null {
-  if (!origin) {
-    return null;
-  }
-
-  const trimmed = origin.trim();
-  if (trimmed === "") {
-    return null;
-  }
-
-  try {
-    return new URL(trimmed).origin;
-  } catch {
-    return trimmed.replace(/\/+$/, "");
-  }
 }
 
 function json(status: number, body: unknown, headers?: HeadersInit): Response {
